feat: add dev console dragging, fix weather dragging
The Developer Console can be dragged from its header. Its position is constrained within the browser window. Double clicking on the header bar resets its position.

The weather container's constraining logic has finally been fixed. It behaves identically to the Developer Console.
