+++ src/global.ts
@@ -175,75 +175,6 @@ menu.borderstyleselect.addEventListener('change', () => {
     }
 });
 
-// Weather move toggle listener
-let isMoving: boolean = false;
-
-panel.section.we.addEventListener('click', (e) => {
-    const target = e.target as HTMLElement;
-    
-    match(target.tagName)
-        .with('BUTTON', () => {
-            const buttonElement = target as HTMLButtonElement;
-            match(buttonElement.id)
-                .with('weatherGeoBtn', async () => {
-                    try {
-                        const latlonArray = await getLocation();
-                        menu.weatherlatinput.value = latlonArray[0].toString();
-                        menu.weatherloninput.value = latlonArray[1].toString();
-                        logConsole(`Retrieved geolocation: ${latlonArray}`, 'debug');
-                    } catch (error) {
-                        logConsole(`Failed to get location: ${error}`, 'error');
-                    }
-                })
-                .with('weatherSubmitBtn', () => {
-                    submitWeatherSettings();
-                })
-                .with('weatherStopBtn', () => {
-                    stopWeather();
-                })
-                .with('weatherMoveToggle', () => {
-                    isMoving = buttonElement.classList.contains('active');
-                    weather.container.style.cursor = isMoving ? 'grab' : 'default';
-                    logConsole(`Weather moving toggle set to: ${isMoving}`, 'debug');
-                })                
-                .with('weatherMoveReset', () => {
-                    weather.container.style.left = '';
-                    weather.container.style.top = '';
-                    logConsole('Weather widget position reset...', 'info');
-                })
-                .otherwise(() => {});
-        })
-        .otherwise(() => {});
-});
-
-weather.container.addEventListener('mousedown', (e) => {
-    if (!isMoving) return;
-    
-    weather.container.style.cursor = 'grabbing';
-    logConsole('Weather widget mousedown...', 'info');
-    const startX = e.clientX - weather.container.offsetLeft;
-    const startY = e.clientY - weather.container.offsetTop;
-
-    function onMouseMove(e: { clientX: number; clientY: number; }) {
-        const posX = e.clientX - startX;
-        const posY = e.clientY - startY;
-
-        weather.container.style.left = `${posX}px`;
-        weather.container.style.top = `${posY}px`;
-        logConsole(`Weather widget moving. PosX: ${posX}, PosY: ${posY}`, 'debug');
-    }
-
-    function onMouseUp() {
-        document.removeEventListener('mousemove', onMouseMove);
-        document.removeEventListener('mouseup', onMouseUp);
-        weather.container.style.cursor = 'grab';
-        logConsole('Weather widget mouseup...', 'info');
-    }
-
-    document.addEventListener('mousemove', onMouseMove);
-    document.addEventListener('mouseup', onMouseUp);
-});
-
 // Menu theme listener
 menu.themeradio.forEach((radio) => {
     radio.addEventListener('change', () => {
