+++ src/index.html
@@ -32,7 +32,7 @@
         <!-- Dev console -->
         <div id="dev-console-container">
             <div class="card">
-                <div class="card-header d-flex justify-content-between align-items-center">
+                <div class="card-header d-flex justify-content-between align-items-center" id="dev-console-header">
                     <span data-i18n="devcon.title"></span>
                     <button type="button" class="btn-close" id="dev-console-close" aria-label="Close"></button>
                 </div>
